+++ NexmoTestApp/ViewControllers/Calls/CallViewController.m
@@ -51,6 +51,9 @@
 
 @implementation CallViewController
 
+- (void)call:(nonnull NXMCall *)call didReceive:(nonnull NSString *)dtmf fromCallMember:(nullable NXMCallMember *)callMember {
+    
+}
 - (void)viewDidLoad {
     [super viewDidLoad];
 }